# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Pesisir Project",
        "description": "API Documentation for Pesisir",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/donations": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get list of donations",
                "operationId": "efd463fc838a5d0b3edcc89ecc8374ac",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Donations"
                ],
                "summary": "Create a new donation",
                "operationId": "a83d56537f533c01a1daf0abefdec913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "image",
                                    "method_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 50000
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "method_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Donation"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/donations/{id}": {
            "get": {
                "tags": [
                    "Donations"
                ],
                "summary": "Get donation by ID",
                "operationId": "d70f7a1ae033692540b40f026aa5740e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Donation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Donation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Donations"
                ],
                "summary": "Delete donation",
                "operationId": "ff34df7029ed4a4fc68e6a59674f9621",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Donation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation deleted successfully"
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                }
            }
        },
        "/api/donation-methods": {
            "get": {
                "tags": [
                    "Donation Methods"
                ],
                "summary": "Get all donation methods",
                "operationId": "2a6001cf68d4369c0482062e108a2872",
                "responses": {
                    "200": {
                        "description": "List of donation methods",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DonationMethod"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Donation Methods"
                ],
                "summary": "Create a new donation method",
                "operationId": "770116b118ee8080b28b4e77fcbcaadc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "method",
                                    "number",
                                    "owner"
                                ],
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "example": "BCA"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "owner": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation method created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonationMethod"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/donation-methods/{id}": {
            "get": {
                "tags": [
                    "Donation Methods"
                ],
                "summary": "Get donation method by ID",
                "operationId": "8847f29cdb7007bed4dc891d5323d8f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Donation method ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation method detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonationMethod"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation method not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Donation Methods"
                ],
                "summary": "Update donation method",
                "operationId": "64e3b7706f446cf325d5aaef4d7cbafc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Donation method ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "method": {
                                        "type": "string",
                                        "example": "BRI"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "owner": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donation method updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DonationMethod"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation method not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Donation Methods"
                ],
                "summary": "Delete donation method",
                "operationId": "b9bda0cc8149e9e50d4a94fba61d8a8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Donation method ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation method deleted"
                    },
                    "404": {
                        "description": "Donation method not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Donation": {
                "title": "Donation",
                "description": "Donation model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "donation_amount": {
                        "type": "integer",
                        "example": 50000
                    },
                    "image_slip": {
                        "type": "string",
                        "example": "donations/slip1.png"
                    },
                    "donation_method_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-23T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-23T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "DonationMethod": {
                "title": "Donation Method",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "method_name": {
                        "type": "string",
                        "example": "BCA"
                    },
                    "account_number": {
                        "type": "string",
                        "example": "1234567890"
                    },
                    "owner_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}